Fix incorrect types in sklearn gp stubs

+++ src/stubs/sklearn/gaussian_process/__init__.pyi
@@ -1,9 +1,11 @@
 from __future__ import annotations
 
 from collections.abc import Callable, Iterable, Sequence
-from typing import Literal, TypeAlias
+from typing import Literal
 
+from numpy import float_
 from numpy.typing import NDArray, ArrayLike
+from typing_extensions import TypeAlias
 
 from sklearn.gaussian_process.kernels import Kernel
 
@@ -14,7 +16,7 @@ class GaussianProcessRegressor:
     def __init__(
         self,
         kernel: Kernel = ...,
-        alpha: float | NDArray[float] = ...,
+        alpha: float | NDArray[float_] = ...,
         optimizer: Literal["fmin_l_bfgs_b"] | _Optimizer | None = ...,
         n_restarts_optimizer: int = ...,
         normalize_y: bool = ...,
@@ -28,5 +30,5 @@ class GaussianProcessRegressor:
         X: ArrayLike,
         return_std: bool = ...,
         return_cov: bool = ...,
-    ) -> tuple[NDArray[float], NDArray[float], NDArray[float]]:
+    ) -> tuple[NDArray[float_], NDArray[float_], NDArray[float_]]:
         ...
